web/fix: resolving conflicts
